I placed the add(+) button on the bottom right of the main screen. Upon clicking the add(+) button, 2 choices will pop up, either 'Reminder' and 'Focus Task'. The difference between the 'Reminder' and 'Focus Task' are mostly on their notifications. 'Reminder' only alarm once, (best used for setting events etc), while 'Focus Task' notifies the user twice (best used for setting a reminder for an Activity). 1st alarm indicates that you should start and the second alarm means the end of your activity (in other words, you need to stop), plus, you can also adjust the time (example +15 minutes).
The hamburger icon - when clicked, a slide screen from the bottom of the main screen will show up consisting 'History' 'Setting' and 'About'. When 'History' is clicked, 12 tiles representing 12 months are there. Each month can be clicked, inside it are the done tasks.
- Done tasks are stored in the history (deleted tasks do not count); only ones that notifies the user will be stored there).

In the main screen, only the ongoing tasks within the particular day will be displayed. Border color depends on the level of the urgency. Red for High Yellow for Medium and Green for Low. Calendar Icon on the top right - when clicked, a calendar will show up. Each days can be clicked and when it does, the ongoing tasks will show on the mid-bottom part of the screen.

Next is the Task Bar - when clicked, done tasks are stored there temporarily, hence the progress bar will change depending on the number of done tasks. (Only done tasks / tasks that notifies are stored here), all tasks inside here will be deleted once the day changed.

In the main screen, you can modify the ongoing tasks by clicking it once.